Update haddocks and project structure
- Update project structure by category
- Update haddocks to reflect project structure
