+++ sample-app/Main.hs
@@ -35,8 +35,8 @@ main = startApp App {..}
     view   = viewModel            -- view function
     events = defaultEvents        -- default delegated events
     subs   = []                   -- empty subscription list
-    mountPoint = Nothing          -- mount point for application (Nothing defaults to 'body')
-    logLevel = Off                -- used during prerendering to see if the VDOM and DOM are in sync (only used with `miso` function)
+    mountPoint = Nothing          -- mount point for application (Nothing defaults to *<body>*)
+    logLevel = Off                -- used during prerendering to see if the VDOM and DOM are in sync (only used with @miso@ function)
 
 -- | Updates model, optionally introduces side effects
 updateModel :: Action -> Model -> Effect Action Model
