+++ src/Miso/Mathml/Element.hs
@@ -18,6 +18,6 @@ import           Miso.Html.Types
 import           Miso.String (MisoString)
 import qualified Prelude            as P
 
--- | Used to construct `VNode`'s in `View`
+-- | Used to construct @Node@ in @View@
 nodeMathml :: MisoString -> [Attribute action] -> [View action] -> View action
 nodeMathml = P.flip (node MATHML) P.Nothing
