+++ src/Miso/Subscription/Window.hs
@@ -39,13 +39,11 @@ windowCoordsSub f = \sink -> do
       Just h <- fromJSVal =<< getProp "innerHeight" (Object target)
       sink $ f (h, w)
 
--- | @windowSub eventName decoder toAction@ is a subscription which parallels the
--- attribute handler `on`, providing a subscription to listen to window level events.
+-- | @windowSub eventName decoder toAction@ provides a subscription to listen to window level events.
 windowSub :: MisoString -> Decoder r -> (r -> action) -> Sub action
 windowSub = windowSubWithOptions defaultOptions
 
--- | @windowSubWithOptions options eventName decoder toAction@ is a subscription which parallels the
--- attribute handler `on`, providing a subscription to listen to window level events.
+-- | @windowSubWithOptions options eventName decoder toAction@ provides a subscription to listen to window level events.
 windowSubWithOptions :: Options -> MisoString -> Decoder r -> (r -> action) -> Sub action
 windowSubWithOptions Options{..} eventName Decoder{..} toAction = \sink ->
   windowAddEventListener eventName $ \e -> do
