+++ src/Miso/Svg.hs
@@ -19,7 +19,7 @@
 --  ]
 -- @
 --
--- More information on how to use `miso` is available on GitHub
+-- More information on how to use miso is available on GitHub
 --
 -- <http://github.com/dmjio/miso>
 --
