+++ src/Miso/Types.hs
@@ -21,7 +21,6 @@
 module Miso.Types
   ( -- ** Types
     App              (..)
-  , Transition
   , View             (..)
   , Key              (..)
   , Attribute        (..)
@@ -32,33 +31,21 @@ module Miso.Types
   , ComponentOptions (..)
   -- ** Classes
   , ToView           (..)
+  , ToKey            (..)
   -- ** Functions
   , defaultApp
-  , toKey
   , component
   , embed
   , embedWith
   , getMountPoint
   , componentOptions
-  , mapAction
-  , fromTransition
-  , toTransition
-  , scheduleIO
-  , scheduleIO_
-  , scheduleIOFor_
-  , scheduleSub
   ) where
 
-import           Control.Monad.Trans.Class (lift)
-import           Control.Monad.Trans.State.Strict (StateT(StateT), execStateT, mapStateT)
-import           Control.Monad.Trans.Writer.Strict (Writer, tell, mapWriter, runWriter)
 import           Data.Aeson (Value)
-import           Data.Bifunctor (second, Bifunctor(..))
-import           Data.Foldable (for_)
 import           Data.JSString (JSString)
 import qualified Data.Map.Strict as M
 import           Data.Maybe (fromMaybe)
-import           Data.Proxy
+import           Data.Proxy (Proxy(Proxy))
 import           Data.String (IsString, fromString)
 import qualified Data.Text as T
 import           GHC.TypeLits (KnownSymbol, symbolVal, Symbol)
@@ -66,7 +53,7 @@ import           Language.Javascript.JSaddle (ToJSVal(toJSVal), Object, JSM)
 import           Prelude hiding (null)
 import           Servant.API (HasLink(MkLink, toLink))
 
-import           Miso.Effect
+import           Miso.Effect (Effect, Sub, Sink)
 import           Miso.Event.Types
 import           Miso.String (MisoString, toMisoString)
 import qualified Miso.String as MS
@@ -77,7 +64,7 @@ data App model action = App
   -- ^ initial model
   , update :: action -> model -> Effect action model
   -- ^ Function to update model, optionally providing effects.
-  --   See the 'Transition' monad for succinctly expressing model transitions.
+  --   See the @Transition@ monad for succinctly expressing model transitions.
   , view :: model -> View action
   -- ^ Function to draw `View`
   , subs :: [ Sub action ]
@@ -120,96 +107,7 @@ data LogLevel
   | DebugPrerender
   deriving (Show, Eq)
 
--- | A monad for succinctly expressing model transitions in the 'update' function.
---
--- @Transition@ is a state monad so it abstracts over manually passing the model
--- around. It's also a writer monad where the accumulator is a list of scheduled
--- IO actions. Multiple actions can be scheduled using
--- @Control.Monad.Writer.Class.tell@ from the @mtl@ library and a single action
--- can be scheduled using 'scheduleIO'.
---
--- Tip: use the @Transition@ monad in combination with the stateful
--- <http://hackage.haskell.org/package/lens-4.15.4/docs/Control-Lens-Operators.html lens>
--- operators (all operators ending in "@=@"). The following example assumes
--- the lenses @field1@, @counter@ and @field2@ are in scope and that the
--- @LambdaCase@ language extension is enabled:
---
--- @
--- myApp = App
---   { update = 'fromTransition' . \\case
---       MyAction1 -> do
---         field1 .= value1
---         counter += 1
---       MyAction2 -> do
---         field2 %= f
---         scheduleIO $ do
---           putStrLn \"Hello\"
---           putStrLn \"World!\"
---   , ...
---   }
--- @
-type Transition action model = StateT model (Writer [Sub action])
-
--- | Turn a transition that schedules subscriptions that consume
--- actions of type @a@ into a transition that schedules subscriptions
--- that consume actions of type @b@ using the supplied function of
--- type @a -> b@.
-mapAction :: (actionA -> actionB) -> Transition actionA model r -> Transition actionB model r
-mapAction = mapStateT . mapWriter . second . fmap . mapSub
-
--- | Convert a @Transition@ computation to a function that can be given to 'update'.
-fromTransition
-    :: Transition action model ()
-    -> (model -> Effect action model) -- ^ model 'update' function.
-fromTransition act = \m ->
-  case runWriter (execStateT act m) of
-    (n, actions) -> Effect n actions
-
--- | Convert an 'update' function to a @Transition@ computation.
-toTransition
-    :: (model -> Effect action model) -- ^ model 'update' function
-    -> Transition action model ()
-toTransition f =
-  StateT $ \m ->
-    case f m of
-      Effect n actions -> do
-        tell actions
-        pure ((), n)
-
--- | Schedule a single IO action for later execution.
---
--- Note that multiple IO action can be scheduled using
--- @Control.Monad.Writer.Class.tell@ from the @mtl@ library.
-scheduleIO :: JSM action -> Transition action model ()
-scheduleIO ioAction = scheduleSub $ \sink -> ioAction >>= sink
-
--- | Like 'scheduleIO' but doesn't cause an action to be dispatched to
--- the 'update' function.
---
--- This is handy for scheduling IO computations where you don't care
--- about their results or when they complete.
-scheduleIO_ :: JSM () -> Transition action model ()
-scheduleIO_ ioAction = scheduleSub $ \_sink -> ioAction
-
--- | Like `scheduleIO_` but generalized to any instance of `Foldable`
---
--- This is handy for scheduling IO computations that return a `Maybe` value
-scheduleIOFor_ :: Foldable f => JSM (f action) -> Transition action model ()
-scheduleIOFor_ io = scheduleSub $ \sink -> io >>= flip for_ sink
-
--- | Like 'scheduleIO' but schedules a subscription which is an IO
--- computation that has access to a 'Sink' which can be used to
--- asynchronously dispatch actions to the 'update' function.
---
--- A use-case is scheduling an IO computation which creates a
--- 3rd-party JS widget which has an associated callback. The callback
--- can then call the sink to turn events into actions. To do this
--- without accessing a sink requires going via a @'Sub'scription@
--- which introduces a leaky-abstraction.
-scheduleSub :: Sub action -> Transition action model ()
-scheduleSub sub = lift $ tell [ sub ]
-
--- | Core type for constructing a `VTree`, use this instead of `VTree` directly.
+-- | Core type for constructing a virtual DOM in Haskell
 data View action
   = Node NS MisoString (Maybe Key) [Attribute action] [View action]
   | Text MisoString
@@ -218,7 +116,7 @@ data View action
   deriving Functor
 
 -- | Options for Components, used with @embedWith@
--- Components are implemented as `div`.
+-- Components are implemented as 'div'.
 data ComponentOptions action
   = ComponentOptions
   { onMounted :: Maybe action
@@ -312,34 +210,38 @@ instance ToJSVal NS where
 -- invariant gives undefined behavior at runtime.
 newtype Key = Key MisoString
 
-instance ToJSVal Key where toJSVal (Key x) = toJSVal x
+-- | ToJSVal instance for Key
+instance ToJSVal Key where
+  toJSVal (Key x) = toJSVal x
 
--- | Convert custom key types to `Key`.
+-- | Convert custom key types to @Key@.
 --
 -- Instances of this class do not have to guarantee uniqueness of the
--- generated keys, it is up to the user to do so. `toKey` must be an
+-- generated keys, it is up to the user to do so. @toKey@ must be an
 -- injective function.
-class ToKey key where toKey :: key -> Key
+class ToKey key where
+  -- | Converts any key into @Key@
+  toKey :: key -> Key
 -- | Identity instance
 instance ToKey Key where toKey = id
--- | Convert `MisoString` to `Key`
+-- | Convert @MisoString@ to @Key@
 instance ToKey JSString where toKey = Key . toMisoString
--- | Convert `T.Text` to `Key`
+-- | Convert @T.Text@ to @Key@
 instance ToKey T.Text where toKey = Key . toMisoString
--- | Convert `String` to `Key`
+-- | Convert @String@ to @Key@
 instance ToKey String where toKey = Key . toMisoString
--- | Convert `Int` to `Key`
+-- | Convert @Int@ to @Key@
 instance ToKey Int where toKey = Key . toMisoString
--- | Convert `Double` to `Key`
+-- | Convert @Double@ to @Key@
 instance ToKey Double where toKey = Key . toMisoString
--- | Convert `Float` to `Key`
+-- | Convert @Float@ to @Key@
 instance ToKey Float where toKey = Key . toMisoString
--- | Convert `Word` to `Key`
+-- | Convert @Word@ to @Key@
 instance ToKey Word where toKey = Key . toMisoString
 
--- | Attribute of a vnode in a `View`.
+-- | Attribute of a vnode in a @View@.
 --
--- The 'Sink' callback can be used to dispatch actions which are fed back to
+-- The @Sink@ callback can be used to dispatch actions which are fed back to
 -- the @update@ function. This is especially useful for event handlers
 -- like the @onclick@ attribute. The second argument represents the
 -- vnode the attribute is attached to.
@@ -349,6 +251,6 @@ data Attribute action
   | S (M.Map MisoString MisoString)
   deriving Functor
 
--- | `IsString` instance
+-- | @IsString@ instance
 instance IsString (View a) where
   fromString = Text . fromString
